2008-01-09  Jeffrey Stedfast  <fejj@novell.com>
	* geometry.cpp (path_get_bounds): My fix the other day wasn't
	quite ideal. Instead of using a 0.1 line thickness when the
	shape's StrokeThickness is 0, continue using that 0 thickness but
	use cairo_fill_extents() instead. Also, when shape is null, ise a
	0.0 thickness & cairo_fill_extents() as well.

	* shape.cpp (DoDraw): Removed the if (thickness == 0) optimization
	that breaks Paths with a StrokeThickness of 0. Fixes bug #352188.

	* runtime.cpp (EmitEventOnList): Changed to take double x,y
	coordinate arguments rather than int.
	(HandleMouseEvent): Same.


svn path=/trunk/moon/; revision=92544

+++ src/runtime.h
@@ -248,9 +248,9 @@ private:
 
 	void FindFirstCommonElement (List *l1, int *index1,
 				     List *l2, int *index2);
-	void EmitEventOnList (MoonlightEventEmitFunc emitter, List *list, int state, int x, int y, int end_idx);
+	void EmitEventOnList (MoonlightEventEmitFunc emitter, List *list, int state, double x, double y, int end_idx);
 	void UpdateCursorFromInputList ();
-	void HandleMouseEvent (MoonlightEventEmitFunc emitter, bool emit_leave, bool emit_enter, bool force_emit, int state, int x, int y);
+	void HandleMouseEvent (MoonlightEventEmitFunc emitter, bool emit_leave, bool emit_enter, bool force_emit, int state, double x, double y);
 	void PerformCapture (UIElement *capture);
 	void PerformReleaseCapture ();
